feat: adjusting index.ts for exporting and adding core.ts
